add magic link base resolver
