+++ src/examples/rest/src/backend/schema/tag/entity.ts
@@ -24,7 +24,7 @@ const acl: AccessControlList<Tag, AuthorizationContext> = {
 @ApplyMultiFactorAuthentication<Tag>({
 	Everyone: {
 		// all users must provide a magic link mfa when writing data
-		Write: [{ factorsRequired: 1, providers: [AuthenticationMethod.PASSKEY] }],
+		Write: [{ factorsRequired: 1, providers: [AuthenticationMethod.MAGIC_LINK] }],
 	},
 })
 @ApplyAccessControlList(acl)
