+++ src/packages/auth/src/authentication/resolvers/magic-link/index.ts
@@ -1,1 +1,2 @@
+export * from './base-resolver';
 export * from './resolver';
